Set times to utc. Skip compiled_version.txt delete after upload. LoopCounter visual

+++ Tritium/dump_uploader.sh
@@ -3,10 +3,12 @@ set -e
 
 OAUTH_ACCESS_TOKEN="<Your dropbox API token goes here>"
 
-uploaddate="$(date +%y-%m-%d)"
-uploadtime="$(date +%H:%M:%S)"
+uploaddate="$(date --utc +%y-%m-%d)"
+uploadtime="$(date --utc +%H:%M:%S)"
 dumpcommit="$(cat ~/.TAO/compiled_version.txt | grep ^commit | awk '{print $2}')"
-
+pushd ~/.TAO/LLL-TAO/
+branch="$(git branch | awk '{print $2}')"
+popd
 [[ ! -f dropbox_uploader.sh ]] && {
 wget https://raw.githubusercontent.com/andreafabrizi/Dropbox-Uploader/master/dropbox_uploader.sh
 }
@@ -20,7 +22,7 @@ debug.log \
 $(ls -1 ~/.TAO/ | grep ^core)
 do \
 [[ -f ~/.TAO/${i} ]] && {
-bash dropbox_uploader.sh upload ~/.TAO/${i} dumps/${dumpcommit}/${uploaddate}-$(hostname)-${uploadtime}-${i}
-rm ~/.TAO/${i}
+bash dropbox_uploader.sh upload ~/.TAO/${i} dumps/${branch}/${uploaddate}/commit-${dumpcommit}/$(hostname)-${uploadtime}-${i}
+[[ "${i}" = "compiled_version.txt" ]] || rm ~/.TAO/${i}
 }
 done
